Added a demo class
This commit is just an example of what the Spaceship class and inner Missile class might look like.

I highly highly recommend we never use this code. It was written before I took a look at the documentation for the given Polygon class. The idea of what I thought Polygon was and what it actually is was not lined up.

In short: Don't use this code. It sucks. It's just an example of what the Spaceship class might look like in future.